+++ CHANGELOG.rst
@@ -26,12 +26,16 @@ Added
 
 - New EnlargeWithAncilla pass for adding ancilla qubits after a Layout
   selection pass (#1603).
-- New Unroll2Q pass for unrolling gates down to just 1q or 2q gates (#1614). 
-- Added a RunConfig object for configurations for run configurations to be used 
+- New Unroll2Q pass for unrolling gates down to just 1q or 2q gates (#1614).
+- Added a RunConfig object for configurations for run configurations to be used
   in compile and circuits_to_qobj. (#1629)
 - Added support for register slicing when applying operations to a register (#1643).
+- Added in new parameter ``justify`` to the text circuit drawer to say how the
+  circuit should be aligned. (#1725)
 - Added function for purity of a mixed state in ``qiskit.quantum_information``
   (#1733)
+- Added parameter to the TextProgressBar to allow the output to be sent to a
+  different output stream 
 
 Changed
 -------
@@ -57,6 +61,17 @@ Changed
   final measurements when compiling for both simulators and devices (#1591).
 - Purity function in ``qiskit.tools.qi.qi`` calls new version in
   ``qiskit.quantum_information`` and issues deprecation warning (#1733)
+- Updated `dag.node_counter` to return the current number of nodes (#1763)
+- The argument ``basis_gates`` used in ``compile``, ``execute``, and ``transpile``
+  is not longer a comma-separated string but a list of strings. For example,
+  this basis ``['u1','u2','u3','cx']`` should be used instead of ``'u1,u2,u3,cx'``
+  (#1333)
+
+Deprecated
+----------
+
+- The methods prefixed by `_get` in the DAGCircuit object are being renamed
+  without that prefix (see #1346)
 
 Fixed
 -----
@@ -68,6 +83,8 @@ Fixed
 - Fixed a bug in BasicMapper pass operating over multiple registers (#1611).
 - Fixed a bug in BarrierBeforeFinalMeasurements which incorrectly moved measurements
   used in conditional operations (#1705).
+- Fixed a bug that with transpile ignoring initial layout when
+  coupling map is provided (#1711).
 
 Removed
 -------